+++ CHANGELOG.md
@@ -105,3 +105,4 @@
 * New Features
     * Added support for nokogiri >= 1.5.1 while maintaining backwards compatibility for >=1.4.2
     * Backports requirement is now for >= 2.3.0
+    * Added the ability to pass in :rest_options to the service constructor within the options hash.